SpreadsheetExpressionEvaluationContextTesting2 extracted SpreadsheetExpressionEvaluationContextTesting
