+++ src/main/java/walkingkooka/spreadsheet/expression/SpreadsheetExpressionEvaluationContextTesting.java
@@ -17,134 +17,43 @@
 
 package walkingkooka.spreadsheet.expression;
 
-import org.junit.jupiter.api.Test;
 import walkingkooka.collect.set.Sets;
-import walkingkooka.spreadsheet.SpreadsheetStrings;
-import walkingkooka.spreadsheet.engine.SpreadsheetDelta;
-import walkingkooka.spreadsheet.environment.SpreadsheetEnvironmentContextTesting2;
+import walkingkooka.spreadsheet.environment.SpreadsheetEnvironmentContextTesting;
 import walkingkooka.spreadsheet.formula.parser.SpreadsheetFormulaParserToken;
-import walkingkooka.spreadsheet.meta.SpreadsheetId;
-import walkingkooka.spreadsheet.meta.SpreadsheetMetadata;
-import walkingkooka.spreadsheet.meta.SpreadsheetMetadataContextTesting2;
-import walkingkooka.spreadsheet.meta.SpreadsheetMetadataPropertyName;
-import walkingkooka.spreadsheet.provider.SpreadsheetProviderContextTesting2;
+import walkingkooka.spreadsheet.meta.SpreadsheetMetadataContextTesting;
+import walkingkooka.spreadsheet.provider.SpreadsheetProviderContextTesting;
 import walkingkooka.spreadsheet.reference.SpreadsheetCellRangeReference;
 import walkingkooka.spreadsheet.reference.SpreadsheetCellReference;
 import walkingkooka.spreadsheet.reference.SpreadsheetColumnReference;
 import walkingkooka.spreadsheet.reference.SpreadsheetLabelMapping;
 import walkingkooka.spreadsheet.reference.SpreadsheetLabelName;
-import walkingkooka.spreadsheet.reference.SpreadsheetLabelNameResolverTesting2;
+import walkingkooka.spreadsheet.reference.SpreadsheetLabelNameResolverTesting;
 import walkingkooka.spreadsheet.reference.SpreadsheetRowReference;
-import walkingkooka.spreadsheet.validation.SpreadsheetValidationReference;
 import walkingkooka.spreadsheet.value.HasSpreadsheetCellTesting;
 import walkingkooka.spreadsheet.value.SpreadsheetCell;
-import walkingkooka.spreadsheet.value.SpreadsheetErrorKind;
-import walkingkooka.storage.expression.function.StorageExpressionEvaluationContextTesting2;
+import walkingkooka.storage.expression.function.StorageExpressionEvaluationContextTesting;
 import walkingkooka.text.CharSequences;
 import walkingkooka.text.cursor.TextCursors;
-import walkingkooka.tree.json.expression.JsonNodeExpressionEvaluationContextTesting2;
-import walkingkooka.validation.expression.ValidatorExpressionEvaluationContextTesting2;
-import walkingkooka.validation.form.expression.FormHandlerExpressionEvaluationContextTesting2;
+import walkingkooka.tree.json.expression.JsonNodeExpressionEvaluationContextTesting;
+import walkingkooka.validation.expression.ValidatorExpressionEvaluationContextTesting;
+import walkingkooka.validation.form.expression.FormHandlerExpressionEvaluationContextTesting;
 
 import java.util.Optional;
 import java.util.Set;
 
-import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public interface SpreadsheetExpressionEvaluationContextTesting<C extends SpreadsheetExpressionEvaluationContext> extends FormHandlerExpressionEvaluationContextTesting2<SpreadsheetValidationReference, SpreadsheetDelta, C>,
+public interface SpreadsheetExpressionEvaluationContextTesting extends FormHandlerExpressionEvaluationContextTesting,
     HasSpreadsheetCellTesting,
-    JsonNodeExpressionEvaluationContextTesting2<C>,
-    SpreadsheetEnvironmentContextTesting2<C>,
-    SpreadsheetLabelNameResolverTesting2<C>,
-    SpreadsheetMetadataContextTesting2<C>,
-    StorageExpressionEvaluationContextTesting2<C>,
-    ValidatorExpressionEvaluationContextTesting2<SpreadsheetValidationReference, C>,
-    SpreadsheetProviderContextTesting2<C> {
-
-    // stringEqualsCaseSensitivity......................................................................................
-
-    @Test
-    default void testStringEqualsCaseSensitivity() {
-        this.checkEquals(
-            SpreadsheetStrings.CASE_SENSITIVITY,
-            this.createContext()
-                .stringEqualsCaseSensitivity()
-        );
-    }
-
-    // evaluate.........................................................................................................
-
-    @Test
-    default void testEvaluateWithEmptyStringReturnsError() {
-        this.evaluateAndCheck(
-            this.createContext(),
-            "",
-            SpreadsheetErrorKind.ERROR.setMessage("End of text, expected \"\\\'\", [STRING] | EQUALS_EXPRESSION | VALUE")
-        );
-    }
-
-    @Test
-    default void testEvaluateWithWhitespaceStringReturnsError() {
-        this.evaluateAndCheck(
-            this.createContext(),
-            " ",
-            SpreadsheetErrorKind.ERROR.setMessage("Invalid character \' \' expected \"\\\'\", [STRING] | EQUALS_EXPRESSION | VALUE")
-        );
-    }
+    JsonNodeExpressionEvaluationContextTesting,
+    SpreadsheetEnvironmentContextTesting,
+    SpreadsheetLabelNameResolverTesting,
+    SpreadsheetMetadataContextTesting,
+    StorageExpressionEvaluationContextTesting,
+    ValidatorExpressionEvaluationContextTesting,
+    SpreadsheetProviderContextTesting {
 
     // parseExpression..................................................................................................
-
-    @Test
-    default void testParseExpressionNullFails() {
-        assertThrows(
-            NullPointerException.class,
-            () -> this.createContext()
-                .parseExpression(null)
-        );
-    }
-
-    @Test
-    default void testParseExpressionWithEmptyStringFails() {
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> this.createContext()
-                .parseExpression(
-                    TextCursors.charSequence("")
-                )
-        );
-    }
-
-    @Test
-    default void testParseExpressionWithOnlyWhitespaceStringFails() {
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> this.createContext()
-                .parseExpression(
-                    TextCursors.charSequence(" ")
-                )
-        );
-    }
-
-    @Test
-    default void testParseExpressionWithOnlyWhitespaceStringFails2() {
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> this.createContext()
-                .parseExpression(
-                    TextCursors.charSequence("  ")
-                )
-        );
-    }
-
-    default void parseExpressionAndCheck(final String formula,
-                                         final SpreadsheetFormulaParserToken expected) {
-        this.parseExpressionAndCheck(
-            this.createContext(),
-            formula,
-            expected
-        );
-    }
 
     default void parseExpressionAndCheck(final SpreadsheetExpressionEvaluationContext context,
                                          final String expression,
@@ -155,15 +64,6 @@ public interface SpreadsheetExpressionEvaluationContextTesting<C extends Spreads
                 TextCursors.charSequence(expression)
             ),
             () -> "parseExpression " + CharSequences.quoteAndEscape(expression) + " with context " + context);
-    }
-
-    default void parseExpressionAndFail(final String expression,
-                                        final String expected) {
-        this.parseExpressionAndFail(
-            this.createContext(),
-            expression,
-            expected
-        );
     }
 
     default void parseExpressionAndFail(final SpreadsheetExpressionEvaluationContext context,
@@ -184,57 +84,6 @@ public interface SpreadsheetExpressionEvaluationContextTesting<C extends Spreads
 
     // parseValueOrExpression...........................................................................................
 
-    @Test
-    default void testParseValueOrExpressionNullFails() {
-        assertThrows(
-            NullPointerException.class,
-            () -> this.createContext()
-                .parseValueOrExpression(null)
-        );
-    }
-
-    @Test
-    default void testParseValueOrExpressionWithEmptyStringFails() {
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> this.createContext()
-                .parseValueOrExpression(
-                    TextCursors.charSequence("")
-                )
-        );
-    }
-
-    @Test
-    default void testParseValueOrExpressionWithOnlyWhitespaceStringFails() {
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> this.createContext()
-                .parseValueOrExpression(
-                    TextCursors.charSequence(" ")
-                )
-        );
-    }
-
-    @Test
-    default void testParseValueOrExpressionWithOnlyWhitespaceStringFails2() {
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> this.createContext()
-                .parseValueOrExpression(
-                    TextCursors.charSequence("  ")
-                )
-        );
-    }
-    
-    default void parseValueOrExpressionAndCheck(final String valueOrExpression,
-                                                final SpreadsheetFormulaParserToken expected) {
-        this.parseValueOrExpressionAndCheck(
-            this.createContext(),
-            valueOrExpression,
-            expected
-        );
-    }
-
     default void parseValueOrExpressionAndCheck(final SpreadsheetExpressionEvaluationContext context,
                                                 final String valueOrExpression,
                                                 final SpreadsheetFormulaParserToken expected) {
@@ -244,15 +93,6 @@ public interface SpreadsheetExpressionEvaluationContextTesting<C extends Spreads
                 TextCursors.charSequence(valueOrExpression)
             ),
             () -> "parseValueOrExpression " + CharSequences.quoteAndEscape(valueOrExpression) + " with context " + context);
-    }
-
-    default void parseValueOrExpressionAndFail(final String valueOrExpression,
-                                               final String expected) {
-        this.parseValueOrExpressionAndFail(
-            this.createContext(),
-            valueOrExpression,
-            expected
-        );
     }
 
     default void parseValueOrExpressionAndFail(final SpreadsheetExpressionEvaluationContext context,
@@ -273,15 +113,7 @@ public interface SpreadsheetExpressionEvaluationContextTesting<C extends Spreads
 
     // loadCell.........................................................................................................
 
-    @Test
-    default void testLoadCellWithNullCellFails() {
-        assertThrows(
-            NullPointerException.class,
-            () -> this.createContext().loadCell(null)
-        );
-    }
-
-    default void loadCellAndCheck(final C context,
+    default void loadCellAndCheck(final SpreadsheetExpressionEvaluationContext context,
                                   final SpreadsheetCellReference cellReference) {
         this.loadCellAndCheck(
             context,
@@ -290,7 +122,7 @@ public interface SpreadsheetExpressionEvaluationContextTesting<C extends Spreads
         );
     }
 
-    default void loadCellAndCheck(final C context,
+    default void loadCellAndCheck(final SpreadsheetExpressionEvaluationContext context,
                                   final SpreadsheetCellReference cellReference,
                                   final SpreadsheetCell expected) {
         this.loadCellAndCheck(
@@ -300,7 +132,7 @@ public interface SpreadsheetExpressionEvaluationContextTesting<C extends Spreads
         );
     }
 
-    default void loadCellAndCheck(final C context,
+    default void loadCellAndCheck(final SpreadsheetExpressionEvaluationContext context,
                                   final SpreadsheetCellReference cellReference,
                                   final Optional<SpreadsheetCell> expected) {
         this.checkEquals(
@@ -312,16 +144,7 @@ public interface SpreadsheetExpressionEvaluationContextTesting<C extends Spreads
 
     // loadCellRange....................................................................................................
 
-    @Test
-    default void testLoadCellRangeWithNullRangeFails() {
-        assertThrows(
-            NullPointerException.class,
-            () -> this.createContext()
-                .loadCellRange(null)
-        );
-    }
-
-    default void loadCellRangeAndCheck(final C context,
+    default void loadCellRangeAndCheck(final SpreadsheetExpressionEvaluationContext context,
                                        final SpreadsheetCellRangeReference range,
                                        final SpreadsheetCell... expected) {
         this.loadCellRangeAndCheck(
@@ -331,7 +154,7 @@ public interface SpreadsheetExpressionEvaluationContextTesting<C extends Spreads
         );
     }
 
-    default void loadCellRangeAndCheck(final C context,
+    default void loadCellRangeAndCheck(final SpreadsheetExpressionEvaluationContext context,
                                        final SpreadsheetCellRangeReference range,
                                        final Set<SpreadsheetCell> expected) {
         this.checkEquals(
@@ -343,15 +166,7 @@ public interface SpreadsheetExpressionEvaluationContextTesting<C extends Spreads
 
     // loadLabel........................................................................................................
 
-    @Test
-    default void testLoadLabelWithNullLabelFails() {
-        assertThrows(
-            NullPointerException.class,
-            () -> this.createContext().loadLabel(null)
-        );
-    }
-
-    default void loadLabelAndCheck(final C context,
+    default void loadLabelAndCheck(final SpreadsheetExpressionEvaluationContext context,
                                    final SpreadsheetLabelName labelName) {
         this.loadLabelAndCheck(
             context,
@@ -360,7 +175,7 @@ public interface SpreadsheetExpressionEvaluationContextTesting<C extends Spreads
         );
     }
 
-    default void loadLabelAndCheck(final C context,
+    default void loadLabelAndCheck(final SpreadsheetExpressionEvaluationContext context,
                                    final SpreadsheetLabelName labelName,
                                    final SpreadsheetLabelMapping expected) {
         this.loadLabelAndCheck(
@@ -370,7 +185,7 @@ public interface SpreadsheetExpressionEvaluationContextTesting<C extends Spreads
         );
     }
 
-    default void loadLabelAndCheck(final C context,
+    default void loadLabelAndCheck(final SpreadsheetExpressionEvaluationContext context,
                                    final SpreadsheetLabelName labelName,
                                    final Optional<SpreadsheetLabelMapping> expected) {
         this.checkEquals(
@@ -382,63 +197,12 @@ public interface SpreadsheetExpressionEvaluationContextTesting<C extends Spreads
 
     // setCell..........................................................................................................
 
-    @Test
-    default void testSetCellWithNullCellFails() {
-        assertThrows(
-            NullPointerException.class,
-            () -> this.createContext()
-                .setCell(null)
-        );
-    }
-
-    @Test
-    default void testSetCellWithSame() {
-        final C context = this.createContext();
-
-        assertSame(
-            context,
-            context.setCell(
-                context.cell()
-            )
-        );
-    }
-
-    default void setCellAndCheck(final C context,
+    default void setCellAndCheck(final SpreadsheetExpressionEvaluationContext context,
                                  final Optional<SpreadsheetCell> cell,
                                  final SpreadsheetExpressionEvaluationContext expected) {
         this.checkEquals(
             expected,
             context.setCell(cell)
-        );
-    }
-
-    // setSpreadsheetMetadata...........................................................................................
-
-    @Test
-    default void testSetSpreadsheetMetadataWithNullFails() {
-        assertThrows(
-            NullPointerException.class,
-            () -> this.createContext()
-                .setSpreadsheetMetadata(null)
-        );
-    }
-
-    @Test
-    default void testSetSpreadsheetMetadataWithDifferentIdFails() {
-        final C context = this.createContext();
-        final SpreadsheetMetadata metadata = context.spreadsheetMetadata();
-        final SpreadsheetMetadata set = metadata.set(
-            SpreadsheetMetadataPropertyName.SPREADSHEET_ID,
-            SpreadsheetId.with(
-                1L +
-                    metadata.getOrFail(SpreadsheetMetadataPropertyName.SPREADSHEET_ID)
-                        .value()
-            )
-        );
-
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> context.setSpreadsheetMetadata(set)
         );
     }
 
@@ -500,37 +264,5 @@ public interface SpreadsheetExpressionEvaluationContextTesting<C extends Spreads
             expected,
             context.nextEmptyRow(column)
         );
-    }
-
-    // spreadsheetFormatterContext......................................................................................
-
-    @Test
-    default void testSpreadsheetFormatterContextWithNullFails() {
-        assertThrows(
-            NullPointerException.class,
-            () -> this.createContext()
-                .spreadsheetFormatterContext(null)
-        );
-    }
-
-    // SpreadsheetLabelNameResolverTesting2..............................................................................
-
-    @Override
-    default C createSpreadsheetLabelNameResolver() {
-        return this.createContext();
-    }
-
-    // class............................................................................................................
-
-    @Override
-    default String typeNameSuffix() {
-        return SpreadsheetExpressionEvaluationContext.class.getSimpleName();
-    }
-
-    // SpreadsheetExpressionEvaluationContext...........................................................................
-
-    @Override
-    default C createConverterLike() {
-        return this.createContext();
     }
 }
