+++ src/test/java/walkingkooka/spreadsheet/expression/SpreadsheetExpressionEvaluationContextConverterTest.java
@@ -86,7 +86,7 @@ import java.util.Optional;
 import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public final class SpreadsheetExpressionEvaluationContextConverterTest implements SpreadsheetExpressionEvaluationContextTesting<SpreadsheetExpressionEvaluationContext>,
+public final class SpreadsheetExpressionEvaluationContextConverterTest implements SpreadsheetExpressionEvaluationContextTesting2<SpreadsheetExpressionEvaluationContext>,
     DecimalNumberContextDelegator,
     SpreadsheetMetadataTesting {
